pyproject.toml
[project]
name = "library-service-v2"
version = "0.1.0"
description = "Single-library book service backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.14.0",
  "fastapi>=0.116.0",
  "openai>=1.0.0",
  "paddleocr>=2.9.1",
  "pgvector>=0.3.6",
  "pydantic-settings>=2.7.0",
  "psycopg[binary]>=3.2.9",
  "python-multipart>=0.0.20",
  "pyjwt>=2.10.1",
  "redis>=6.4.0",
  "sqlalchemy>=2.0.37",
  "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
  "httpx>=0.28.1",
  "pytest>=8.3.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
